Allow function to be specified as callback
Backbone events could be written inline.

This is useful for for cases where it doesn't quite make
sense to create a method just for them.  Such as:

Backbone.View.extend({
  events:
     'click button': function(){ this.quack(1) }

  quack: function(number_of_times){

  }

  unrelated_method: function(){
    this.quack(3);
  }
})

If the callback was just 'quack', it would be passed the Event object.

It can also be useful for adding events later in the view's lifecycle by calling
this.eventManager.bind() with a function that has closure scope.

This extends the events so they can be used similarly.